requirements.txt
python-dotenv>=1.0.0
streamlit>=1.32.0
groq>=0.4.2
chromadb>=0.4.0
sentence-transformers>=2.6.0
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
google-generativeai>=0.8.6
groq>=1.1.2
